refactor: rename hooks
