+++ project-4/hooks/index.ts
@@ -1,15 +1,15 @@
 import usePeer from './use-peer';
 import useStream from './use-stream';
-import usePeerOnJoinRoom from './use-peer-on-join-room';
-import usePeerOnAnswer from './use-peer-on-answer';
+import useUserJoin from './use-user-join';
+import useUserAnswer from './use-user-answer';
 import useScreenShare from './use-screen-share';
 import useIsAudioActive from './use-is-audio-active';
 
 export {
   usePeer,
   useStream,
-  usePeerOnJoinRoom,
-  usePeerOnAnswer,
+  useUserJoin,
+  useUserAnswer,
   useScreenShare,
   useIsAudioActive,
 };
